backend/requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.1

# Postgres driver (for Supabase)
psycopg2-binary==2.9.9
